rename `chat_v2` to `chat`
