+++ chat/gacha_reply.go
@@ -1,4 +1,4 @@
-package chat_v2
+package chat
 
 import (
 	"csust-got/config"
