+++ config/chat.go
